twitter-scraper/requirements.txt
Flask==2.0.2
asyncio==3.4.3
twscrape==0.11.1
python-dotenv==0.19.2
Werkzeug==2.2.2